requirements.txt
Flask==1.1.1
Flask-Compress==1.4.0
Flask-MySQLdb==0.2.0
Flask-SeaSurf==0.2.2
gunicorn==20.0.4
Jinja2==2.10.1
mysql==0.0.2
mysqlclient==1.4.6
PyYAML==5.3
Werkzeug==1.0.0
